+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# BetterBlizzPlates 1.9.3
+## Prepatch/Midnight
+### Tweak
+- Tweak Class Indicator position for some settings
+- Update profiles for Midnight: Aeghis, Starter, Blitz, Bodify (Likely more tweaks later on)
+### Bugfix
+- Fix nameplate debuff padding resetting.
+- Fix no-interrupt castbar coloring.
+
 # BetterBlizzPlates 1.9.2g
 ## All versions
 ### Tweak